+++ packages/examples/src/eclipse.jdt.ls/config.ts
@@ -2,7 +2,8 @@
  * Copyright (c) 2024 TypeFox and others.
  * Licensed under the MIT License. See LICENSE in the package root for license information.
  * ------------------------------------------------------------------------------------------ */
-
-import { runGroovyLanguageServer } from './main.js';
-
-runGroovyLanguageServer();
+export const eclipseJdtLsConfig = {
+    port: 30003,
+    path: '/jdtls',
+    basePath: '/home/mlc/packages/examples/resources/eclipse.jdt.ls'
+};