+++ packages/examples/src/eclipse.jdt.ls/server/direct.ts
@@ -3,6 +3,6 @@
  * Licensed under the MIT License. See LICENSE in the package root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-import { runGroovyLanguageServer } from './main.js';
+import { runEclipseJdtLs } from './main.js';
 
-runGroovyLanguageServer();
+runEclipseJdtLs();