+++ packages/examples/src/groovy/client/main.ts
@@ -20,7 +20,7 @@ export const configureMonacoWorkers = () => {
 };
 
 const code = `package test.org;
-import java.io.File ;
+import java.io.File;
 File file = new File("E:/Example.txt");
 `;
 
@@ -44,7 +44,8 @@ const userConfig: UserConfig = {
             userConfiguration: {
                 json: JSON.stringify({
                     'workbench.colorTheme': 'Default Dark Modern',
-                    'editor.guides.bracketPairsHorizontal': 'active'
+                    'editor.guides.bracketPairsHorizontal': 'active',
+                    'editor.wordBasedSuggestions': 'off'
                 })
             }
         }