+++ packages/client/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this npm module are documented in this file.
 
 ## [8.7.0] - 2024-07-1x
 
-- Update to latest monaco-vscode-api [#7xx](https://github.com/TypeFox/monaco-languageclient/pull/7xx)
+- Update to latest monaco-vscode-api [#707](https://github.com/TypeFox/monaco-languageclient/pull/7xx)
   - Updated to version `7.0.3` of `@codingame/monaco-vscode` packages
 
 ## [8.6.0] - 2024-06-29